requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
flask>=3.0
pdfplumber>=0.10
python-docx>=1.1
